requirements.txt
Django>=4.2
djangorestframework>=3.14
textblob>=0.18.0
nltk>=3.9.1